// node-ui/render-process/src/app/local-service-key.enum.ts
// Copyright (c) 2017-2019, Substratum LLC (https://substratum.net) and/or its affiliates. All rights reserved.

export enum LocalServiceKey {
  ChainName = 'chainName',
  NeighborNodeDescriptor = 'neighborNodeDescriptor',
  PersistNeighborPreference = 'persistNeighborPreference',
  BlockchainServiceUrl = 'blockchainServiceUrl',
}